backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
python-dotenv
google-genai